requirements.txt
pyspark==3.5.5
psycopg2
python-dotenv
findspark
boto3